refactor - extracted static class
