+++ src/test/java/beam/analysis/plots/RealizedModeGraphTest.java
@@ -41,7 +41,7 @@ public class RealizedModeGraphTest {
     private Map<Integer, Map<String, Double>> stats;
     private static final String BASE_PATH = Paths.get(".").toAbsolutePath().toString();
     private static final String EVENT_FILE_PATH = BASE_PATH + "/test/input/beamville/test-data/replanning.event.xml";
-    private RealizedModeAnalysis realizedModeStats = new RealizedModeAnalysis(new RealizedModeAnalysis.RealizedModesStatsComputation() {
+    private final RealizedModeAnalysis realizedModeStats = new RealizedModeAnalysis(new RealizedModesStatsComputation() {
         @Override
         public double[][] compute(Tuple<Map<Integer, Map<String, Double>>, Set<String>> stat) {
             stats = stat.getFirst();
